backend/service_register/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
requests==2.30.0
flask_mysqldb
mysql-connector-python==8.0.27

